Add support for Python 3.11 - new class `PyInterpreterFrame` because `PyFrameObject` objects may not be available (in contrast to `PyFrameObject` `PyInterpreterFrame` is not a `PyObject`!) - new command `!pyinterpreterframe` that is used for the `[Frame]` link (We cannot use `!pyobj` anymore because `PyInterpreterFrame` is not a subclass of `PyObject`) - format of mapping between instructions and line numbers changed significantly (again), see `PyCodeObject` - the `__dict__` of Python objects is not stored as `PyDictObject` anymore but as a "managed dict", see `PyObject` and `PyManagedDict` - the symbol `PyMemberDef` is not available (will be available again in 3.12), see `PyTypeObject`, `PyMemberDefAuto` and `PyMemberDefManual` - new class `PyDictKeysObject` because it is used in `PyDictObject` and `PyManagedDict`

+++ src/ExtHelpers.cpp
@@ -7,11 +7,11 @@ using namespace std;
 
 namespace utils {
 
-	auto getPointerSize() -> uint64_t
+	auto getPointerSize() -> int
 	{
 		string objExpression = "sizeof(void*)"s;
 		ExtRemoteTyped remoteObj(objExpression.c_str());
-		return utils::readIntegral<uint64_t>(remoteObj);
+		return utils::readIntegral<int>(remoteObj);
 	}
